engine: extract esm/tours/tour_edit.js
